+++ src/gui/annotator/settings/ItemSettingsWidgetConfigurator.h
@@ -28,6 +28,7 @@
 #include "src/widgets/ToolPicker.h"
 #include "src/widgets/NumberPicker.h"
 #include "src/widgets/StickerPicker.h"
+#include "src/widgets/ToggleButton.h"
 
 namespace kImageAnnotator {
 
@@ -45,6 +46,7 @@ public:
 	void setFirstNumberWidget(NumberPicker *widget);
 	void setObfuscateFactorWidget(NumberPicker *widget);
 	void setStickerWidget(StickerPicker *widget);
+	void setShadowToggleWidget(ToggleButton *widget);
 
 private:
 	Tools mCurrentTool;
@@ -56,6 +58,7 @@ private:
 	NumberPicker *mFirstNumberWidget;
 	NumberPicker *mObfuscateFactorWidget;
 	StickerPicker *mStickerWidget;
+	ToggleButton *mShadowToggleWidget;
 
 	void updateWidgets() const;
 	void updateProperties() const;
@@ -69,6 +72,7 @@ private:
 	void setNoFillAndNoBorderVisible(bool enabled) const;
 	void setObfuscateFactorWidgetVisible(bool enabled) const;
 	void setStickerWidgetVisible(bool enabled) const;
+	void setShadowToggleWidgetVisible(bool enabled) const;
 	void setWidthRange(int min, int max) const;
 };